Refactored test cases to remove duplicate test cases
