+++ composer.json
@@ -10,7 +10,8 @@
     ],
     "autoload": {
         "psr-4": {
-            "HKarlstrom\\OpenApiReader\\": "src"
+            "HKarlstrom\\OpenApiReader\\": "src",
+            "HKarlstrom\\OpenApiReader\\Tests\\": "tests"
         }
     },
     "require": {
